requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
faker>=19.0.0
openai>=1.3.0
langchain>=0.0.350
langchain-openai>=0.0.2
python-dotenv>=1.0.0
plotly>=5.15.0
seaborn>=0.12.0
matplotlib>=3.7.0
numpy>=1.24.0
tiktoken>=0.5.0
pydantic
pyyaml
psutil
